prevent crashing on shadow DOM elements (#857)
# why
- when the LLM returns an ID, we look up that ID in our ID->xpath
mapping. If that element does not exist, it returns undefined.
downstream, we attempt to run some regex on the xpath
- when running the regex on the undefined xpath, we get the following
error:
`TypeError: Cannot read properties of undefined (reading 'replace')`
- the actual fix for this is to add support for taking actions on
elements within the shadow dom
- for now, we should have better error handling for this case
# what changed
- added a check in the `ObserveHandler` to check if the returned ID
contains `-`, which indicates that it will be present in our `idToXpath`
mapping
- if it does not contain `-`, then we log an error, and return
`"not-supported"` as the selector, and `"an element inside a shadow
DOM"` as the description
# test plan
- added an eval to ensure that we are returning `"not-supported"` as the
selector, and `"an element inside a shadow DOM"` for elements that exist
inside the shadow dom

---------

Co-authored-by: Sean McGuire <seanmcguire1@outlook.com>